moved test to its own class. (setup logic got in the way)
